Add files via upload
When you look around, you will observe different animals with different
structures and forms. As over a million species of animals have been
described till now, the need for classification becomes all the more
important. The classification also helps in assigning a systematic position
to newly described species.
4.1 BASIS OF CLASSIFICATION
Inspite of differences in structure and form of different animals, there are
fundamental features common to various individuals in relation to the
arrangement of cells, body symmetry, nature of coelom, patterns of
digestive, circulatory or reproductive systems. These features are used
as the basis of animal classification and some of them are discussed here.